Change to int

+++ backoffice-service/src/main/java/com/tmoreno/mooc/backoffice/course/create/CreateLessonParams.java
@@ -2,6 +2,6 @@ package com.tmoreno.mooc.backoffice.course.create;
 
 public record CreateLessonParams(
     String title,
-    double durationInSeconds
+    int durationInSeconds
 ) {
 }
